+++ src/task-parser.ts
@@ -1,0 +1,2 @@
+// export { TaskPaperNode } from "../lib/task-parser/TaskPaperNode";
+export { TaskPaperNode } from "task-parser/TaskPaperNode";